Updated SIR for multiple low-high catch streams
Removed all premodern catch and updated input to have it be lists of catch streams where each layer is a seperate catch series with either Year, Catch; or Year, MinCatch, MaxCatch as columns. catch_multipliers are used for SLR and catch_parameters are used for theta to sample between min catch and max catch, where applicable.